fixed the playback file not working
the audio from the microphone was being recorded at 32 bits and that didn't play well with pyaudio. I don't know why pyaudio doesn't like it given that simpleaudio should support 32 bit  audio based on what I saw, but who knows. Playback works now and I just need to add code to copy it into the example directory.

+++ engine/sounds.py
@@ -59,7 +59,7 @@ class SoundManager:
         #
         # the dtype='int32' sets the datatype which ensures that the scipy.io writes it as as pcm wav file
         # if you don't include it then the speech recognition won't open the file because it only reads pcm wav files
-        recording = sounddevice.rec(int(record_length*fs), fs, channels=2, dtype='int32')
+        recording = sounddevice.rec(int(record_length*fs), fs, channels=2, dtype='int16')
         sounddevice.wait()
         scipy.io.wavfile.write(output_file_path, fs, recording)
